Split ext/mail out of ext/mailparse
